ENH: refactored python bindings into crate feature
